requirements.txt
typer==0.7.0
Pillow==9.3.0
